requirement.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Click==7.0
colorama==0.4.3
contextlib2==0.6.0
cryptography==2.8
dbus-python==1.2.16
decouple==0.0.7
defusedxml==0.6.0
des==1.0.5
diff-match-patch==20200713
Django==3.1.5
django-admin-rangefilter==0.6.4
django-cors-headers==3.6.0
django-crispy-forms==1.11.0
django-deepzoom==3.0.3
django-filter==2.4.0
django-import-export==2.5.0
django-resized==0.3.11
django-rest-auth==0.9.5
django-utils-six==2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
idna==2.10
isort==5.7.0
jdcal==1.4.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.1.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
rsa==4.6
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
